backend/requirements.txt
fastapi
uvicorn[standard]
apscheduler
supabase
google-genai
python-dotenv
httpx
pydantic
